Swapping Long to BigDecimal (#382)
* Swapped `Long` for `BigDecimal` in places that will need it, specifically for large nhash quantities
* Should fix the Rich List issue
* Now rounding circulating supply, community pool values as those are whole `nhash` values

closes: #380
closes: #381

+++ service/src/main/kotlin/io/provenance/explorer/service/GovService.kt
@@ -284,7 +284,7 @@ class GovService(
                             )
                             .let { valService.getValidatorsByHeight(it + 1) }
                             .validatorsList
-                            .sumOf { it.votingPower }
+                            .sumOf { it.votingPower.toBigDecimal() }
                             // Voting power is in mhash, not hash. So pad up to nhash for FE UI conversion
                             .mhashToNhash()
